French translation improvement
I worked on an improvement for the french translation of your package.

The internal rules (eg. username) benefit from a better explanation in case of errors.

Rules that can be applied on fields that are often handled by neophyte are explained in order to be understable (eg. titlecase, snakecase, ...).

If you agree with this, it could be nice to take care of the other translations in this way, in order to improve the default error messages.